Explicit (uint8_t) for node height
According to kvlds/node.h line 54,
	/* 1 if this node is a root; 0 otherwise or !present. */
	unsigned int root : 1;
and line 68,
	/* Height of this node (leaf = 0); -1 if !present. */
	int8_t height;

Since we're serializing, the node is present (checked with an assert()), so:
	0 <= N->height <= 127

Therefore:
	. if the node is not a root, we take the latter option of
		*p = (uint8_t)N->height;
	  We can store anything from 0 to 127 in (uint8_t), so the cast
	  is safe.

	. if the node is a root, we take the former option of
		*p = (uint8_t)(0x80 + N->height);
	  This results in a value from 128 to 255, which is safe to
	  store in (uint8_t), so the cast is safe.

+++ kvlds/serialize.c
@@ -106,9 +106,9 @@ serialize(struct btree * T, struct node * N, size_t buflen)
 
 	/* Write height and rootedness. */
 	if (N->root)
-		*p = 0x80 + N->height;
+		*p = (uint8_t)(0x80 + N->height);
 	else
-		*p = N->height;
+		*p = (uint8_t)N->height;
 	p += 1;
 
 	/* Write the matching prefix length. */
